add icalendar to download schedule
